requirements.txt
acres==0.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.5.3
asgiref==3.8.1
asyncpg==0.30.0
attrs==24.3.0
azure-ai-documentintelligence==1.0.2
azure-ai-ml==1.27.1
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.33.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-mgmt-core==1.5.0
azure-monitor-opentelemetry==1.6.10
azure-monitor-opentelemetry-exporter==1.0.0b33
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
azure-storage-file-share==12.21.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
ci-info==0.3.0
click==8.1.8
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
docstring_parser==0.16
durationpy==0.9
et_xmlfile==2.0.0
etelemetry==0.3.1
faiss-cpu==1.9.0.post1
fastapi==0.115.12
filelock==3.16.1
filetype==1.2.0
fitz==0.0.1.dev2
fixedint==0.1.6
Flask==3.1.0
Flask-Cors==5.0.0
flask-restx==1.3.0
flatbuffers==24.12.23
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google_search_results==2.4.1
googleapis-common-protos==1.66.0
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.69.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httplib2==0.22.0
httpsx==0.0.1
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.1
kiwisolver==1.4.8
kubernetes==31.0.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.64
langchain-openai==0.3.19
langchain-postgres==0.0.14
langchain-text-splitters==0.3.5
langgraph==0.4.2
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.45
looseversion==1.3.0
lxml==5.4.0
magika==0.6.2
mammoth==1.9.1
markdown-it-py==3.0.0
markdownify==1.1.0
markitdown==0.1.2
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.1
mdurl==0.1.2
mem0ai==0.1.44
mmh3==5.0.1
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.38.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.6
onnxruntime==1.20.1
openai==1.76.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-dbapi==0.52b1
opentelemetry-instrumentation-django==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-instrumentation-flask==0.52b1
opentelemetry-instrumentation-psycopg2==0.52b1
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-urllib==0.52b1
opentelemetry-instrumentation-urllib3==0.52b1
opentelemetry-instrumentation-wsgi==0.52b1
opentelemetry-proto==1.29.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.18
ormsgpack==1.9.1
overrides==7.7.0
packaging==24.2
panda==0.3.1
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20250327
pdfplumber==0.11.6
pgvector==0.3.6
pillow==11.2.1
pluggy==1.5.0
portalocker==2.10.1
posthog==3.8.3
promptflow==1.17.1
promptflow-core==1.17.1
promptflow-devkit==1.17.1
promptflow-tools==1.4.0
promptflow-tracing==1.17.1
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
prov==2.0.1
psutil==5.9.8
psycopg==3.2.9
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
puremagic==1.28
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.7.1
pydantic_core==2.33.1
pydash==7.0.7
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
pyxnat==1.6.3
PyYAML==6.0.2
qdrant-client==1.12.2
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.37
starlette==0.46.2
streamlit==1.45.0
strictyaml==1.7.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traits==7.0.2
transformers==4.51.3
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
waitress==3.0.2
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.2
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
azure-cosmos
